Color Picker Working
I've got the color wheel able to read and select colors! I haven't yet figured out how to get it to change the drawing color, but I have a hunch it might have to do with something called 'Blueprint Interfaces'. I tried casting as well, but that didn't seem to work because neither 'Drawing' nor 'ColorWheel' is an actor.